add directive  chart-dataset-overload
We could not set the options included in the chart dataset. We could only set the global options thanks to the directive chart-options. Therfore we could not set an option  per chart like described in Issue https://github.com/jtblin/angular-chart.js/issues/370 .
I added a directive chart-dataset-overload. If it is present, createChart will merge both the dataset and the data set overload at the final step of the dataset creation.